added buttons to the button-component
