+++ src/components/Button.js
@@ -2,8 +2,6 @@ import React from 'react'
 import { useDispatch } from 'react-redux'
 import { quiz } from 'reducers/quiz'
 
-
-
 export const RestartButton = () => {
 const dispatch = useDispatch()
 
@@ -19,3 +17,26 @@ const clickRestart = () => {
     )
 
 }
+
+export const AnswerButton = ({item, index, questionId}) => {
+    const dispatch = useDispatch()
+    const onAnswerSubmit = (id, index) => {
+        dispatch(quiz.actions.submitAnswer({ questionId: id, answerIndex: index }));
+      };
+      
+    return (
+        <button className=""type="submit" onClick={()=>onAnswerSubmit(questionId, index)} key={item}>{item}</button>
+    )
+}
+
+export const NextQuestion = () => {
+    const dispatch = useDispatch()
+    const nextQuestion = () => {
+        dispatch(quiz.actions.goToNextQuestion());
+    }
+    
+    return (
+    <button type="submit" onClick={()=>nextQuestion ()}>Next question</button>
+    )
+}
+
